Create The digital world in layman's term and identifying a good programmer
Hey @mariibella and @pvaleiras! As mentioned in our previous discussions, I have decided to write how to communicate with a programmer and also point out some ways of identifying a good programmer. Give me a :+1: or an :ok_hand: if you feel that it's good to :ship:. If you have any further :bulb: to make this better, feedback and comments are WELCOME! 

Also, once we're done with the project, let's find time to go out for :beers: to celebrate the rollout!